[REFACTOR] Delete LotteryOnChainService
