+++ src/main/java/com/example/demo/domain/lottery/service/LotteryService.java
@@ -1,9 +1,15 @@
 package com.example.demo.domain.lottery.service;
 
+import com.example.demo.domain.concert.entity.Session;
+
 import java.math.BigInteger;
 import java.util.List;
 
 public interface LotteryService {
+
+    void commitApplicants(Session session);
+
+    void drawWinners(Long sessionId);
 
     void saveWinners(
             Long sessionId, int count, List<BigInteger> indices, String txHash, Long blockNo, Integer logIdx);
